pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "lxml>=5.3.1",
    "numpy>=2.2.3",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pysftp>=0.2.9",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "rarfile>=4.2",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "schedule>=1.2.2",
    "trafilatura>=2.0.0",
    "watchdog>=6.0.0",
    "xlwt>=1.3.0",
]
